requirements-ml-cpu.txt
# ML Requirements for CPU-only testing on Windows
# Patent #4 Implementation - CPU Version

# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning - CPU versions
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2 --index-url https://download.pytorch.org/whl/cpu

# Gradient Boosting
xgboost==1.7.6
lightgbm==4.0.0

# Time Series
prophet==1.1.4
statsmodels==0.14.0

# Model Explanation
shap==0.42.1

# Async and WebSocket
aiohttp==3.8.5
aioredis==2.0.1
websockets==11.0.3

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.11.2

# Azure Integration
azure-identity==1.14.0
azure-mgmt-resource==23.0.1
azure-mgmt-monitor==6.0.1
azure-mgmt-security==3.0.0
azure-mgmt-policyinsights==1.0.0

# API Framework
fastapi==0.101.0
uvicorn[standard]==0.23.2
pydantic==2.1.1
python-multipart==0.0.6

# ML Operations
mlflow==2.5.0
optuna==3.2.0

# Model Serving
onnx==1.14.0
onnxruntime==1.15.1

# Monitoring
prometheus-client==0.17.1

# Security
cryptography==41.0.3
python-jose[cryptography]==3.3.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
jsonschema==4.18.4
click==8.1.6
rich==13.5.2
tqdm==4.65.0

# Redis
redis==4.6.0
hiredis==2.2.3

# Logging
loguru==0.7.0
structlog==23.1.0

# Additional Scientific Computing
networkx==3.1
matplotlib==3.7.2
plotly==5.15.0